Made changes inside right file
