+++ python/fledge/services/core/api/update.py
@@ -11,6 +11,8 @@ import json
 from aiohttp import web
 import datetime
 import os
+import asyncio
+import re
 
 from fledge.services.core import server
 from fledge.common import logger
@@ -27,6 +29,7 @@ __version__ = "${VERSION}"
 _help = """
     -----------------------------------------
     | PUT             | /fledge/update      |
+    | GET             | /fledge/update      |
     -----------------------------------------
 """
 
@@ -107,3 +110,44 @@ async def update_package(request):
         raise web.HTTPInternalServerError(reason=msg, body=json.dumps({"message": msg}))
     else:
         return web.json_response({"status": "Running", "message": status_message})
+
+
+async def get_updates(request: web.Request) -> web.Response:
+    update_cmd = "apt update"
+    update_process = await asyncio.create_subprocess_shell(update_cmd,
+                                                           stdout=asyncio.subprocess.PIPE,
+                                                           stderr=asyncio.subprocess.PIPE)
+
+    _, _ = await update_process.communicate()
+    if update_process.returncode != 0:
+        _logger.error("Could not run {}".format(update_cmd))
+        return web.json_response({'updates': []})
+
+    cmd = "apt list --upgradable | grep \^fledge"
+    installed_packages_process = await asyncio.create_subprocess_shell(cmd,
+                                                                 stdout=asyncio.subprocess.PIPE,
+                                                                 stderr=asyncio.subprocess.PIPE)
+
+    stdout, stderr = await installed_packages_process.communicate()
+    if installed_packages_process.returncode == 0:
+        process_output = stdout.decode("utf-8")
+        _logger.info(process_output)
+        # split on new-line
+        word_list = re.split(r"\n+", process_output)
+
+        # remove '' from the list
+        word_list = [w for w in word_list if w != '']
+        packages = []
+
+        # Now match the character / . The string before / is the actual package name we want.
+        for word in word_list:
+            word_match = re.findall(r".*[/]", word)
+            if len(word_match) > 0:
+                packages.append(word_match[0].replace('/', ''))
+
+        # Make a set to avoid duplicates.
+        upgradable_packages = list(set(packages))
+        return web.json_response({'updates': upgradable_packages})
+    else:
+        _logger.info("Updates are not available at the moment.")
+        return web.json_response({'updates': []})
